# src/routes/file.js
const express = require('express');
const router = express.Router();
const supabase = require('../config/supabase');
const fs = require('fs');
const path = require('path');

// GET all files
router.get('/', async (req, res) => {
  const { data, error } = await supabase.from('file').select('*');
  if (error) return res.status(500).json({ error: error.message });
  res.json(data);
});

// GET file by id
router.get('/:id', async (req, res) => {
  const { id } = req.params;
  const { data, error } = await supabase.from('file').select('*').eq('id', id).single();
  if (error) return res.status(404).json({ error: error.message });
  res.json(data);
});

// POST create file
router.post('/', async (req, res) => {
  try {
    const { filename, mimeType, content } = req.body;
    if (!filename || !content) {
      return res.status(400).json({ error: 'filename and content are required' });
    }
    // Décoder le contenu base64
    const buffer = Buffer.from(content, 'base64');
    // Déterminer le chemin de sauvegarde et gérer l'incrémentation
    const dir = path.join(__dirname, '../../public/file');
    const ext = path.extname(filename);
    const base = path.basename(filename, ext);
    let finalFilename = filename;
    let filePath = path.join(dir, finalFilename);
    let i = 1;
    while (fs.existsSync(filePath)) {
      finalFilename = `${base}(${i})${ext}`;
      filePath = path.join(dir, finalFilename);
      i++;
    }
    // Écrire le fichier sur le disque
    fs.writeFileSync(filePath, buffer);
    // Enregistrer le chemin dans la base (optionnel)
    const dbInput = {
      filename: finalFilename,
      path: `/public/file/${finalFilename}`,
      mimeType,
      created_at: new Date().toISOString()
    };
    const { data, error } = await supabase.from('file').insert(dbInput).select('*').single();
    if (error) return res.status(400).json({ error: error.message });
    res.status(201).json({ ...data, filePath: dbInput.path });
  } catch (err) {
    console.error('Erreur lors de la sauvegarde du fichier:', err);
    res.status(500).json({ error: 'Erreur lors de la sauvegarde du fichier' });
  }
});

// POST create export record
router.post('/export', async (req, res) => {
  try {
    const { file, path } = req.body;
    if (!file || !Array.isArray(path) || path.length === 0) {
      return res.status(400).json({ error: 'file (string) and path (array) are required' });
    }
    const dbInput = {
      file,
      path,
      created_at: new Date().toISOString()
    };
    const { data, error } = await supabase.from('export').insert(dbInput).select('*').single();
    if (error) return res.status(400).json({ error: error.message });
    res.status(201).json(data);
  } catch (err) {
    console.error('Erreur lors de l\'insertion dans export:', err);
    res.status(500).json({ error: 'Erreur lors de l\'insertion dans export' });
  }
});

// PUT update file
router.put('/:id', async (req, res) => {
  const { id } = req.params;
  const input = req.body;
  const { data, error } = await supabase.from('file').update(input).eq('id', id).select('*').single();
  if (error) return res.status(400).json({ error: error.message });
  res.json(data);
});

// DELETE file
router.delete('/:id', async (req, res) => {
  const { id } = req.params;
  const { error } = await supabase.from('file').delete().eq('id', id);
  if (error) return res.status(400).json({ error: error.message });
  res.status(204).send();
});

module.exports = router; 