# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Supabase GraphQL API",
    "description": "API for managing companies, establishments, officers, addresses, files, and related data using GraphQL with a Supabase backend. All operations are performed via a single GraphQL endpoint using POST requests.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local GraphQL server"
    }
  ],
  "paths": {
...
    "/api/establishments": {
      "get": { "summary": "Get all establishments", "tags": ["Establishment"], "responses": { "200": { "description": "List of establishments", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Establishment" } } } } } } },
      "post": { "summary": "Create a new establishment", "tags": ["Establishment"], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EstablishmentInput" } } } }, "responses": { "201": { "description": "Establishment created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Establishment" } } } }, "400": { "description": "Invalid input" } } }
    },
    "/api/establishments/{id}": {
      "get": { "summary": "Get an establishment by ID", "tags": ["Establishment"], "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } } ], "responses": { "200": { "description": "Establishment found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Establishment" } } } }, "404": { "description": "Establishment not found" } } },
      "put": { "summary": "Update an establishment by ID", "tags": ["Establishment"], "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } } ], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EstablishmentInput" } } } }, "responses": { "200": { "description": "Establishment updated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Establishment" } } } }, "400": { "description": "Invalid input" }, "404": { "description": "Establishment not found" } } },
      "delete": { "summary": "Delete an establishment by ID", "tags": ["Establishment"], "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } } ], "responses": { "204": { "description": "Establishment deleted" }, "404": { "description": "Establishment not found" } } }
    },
    "/api/officers": {
      "get": { "summary": "Get all officers", "tags": ["Officer"], "responses": { "200": { "description": "List of officers", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Officer" } } } } } } },
      "post": { "summary": "Create a new officer", "tags": ["Officer"], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OfficerInput" } } } }, "responses": { "201": { "description": "Officer created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Officer" } } } }, "400": { "description": "Invalid input" } } }
    },
    "/api/officers/{id}": {
      "get": { "summary": "Get an officer by ID", "tags": ["Officer"], "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } } ], "responses": { "200": { "description": "Officer found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Officer" } } } }, "404": { "description": "Officer not found" } } },
      "put": { "summary": "Update an officer by ID", "tags": ["Officer"], "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } } ], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OfficerInput" } } } }, "responses": { "200": { "description": "Officer updated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Officer" } } } }, "400": { "description": "Invalid input" }, "404": { "description": "Officer not found" } } },
      "delete": { "summary": "Delete an officer by ID", "tags": ["Officer"], "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } } ], "responses": { "204": { "description": "Officer deleted" }, "404": { "description": "Officer not found" } } }
    },
    "/api/files": {
      "get": { "summary": "Get all files", "tags": ["File"], "responses": { "200": { "description": "List of files", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/File" } } } } } } },
      "post": { "summary": "Create a new file", "tags": ["File"], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FileInput" } } } }, "responses": { "201": { "description": "File created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/File" } } } }, "400": { "description": "Invalid input" } } }
    },
    "/api/files/{id}": {
      "get": { "summary": "Get a file by ID", "tags": ["File"], "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } } ], "responses": { "200": { "description": "File found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/File" } } } }, "404": { "description": "File not found" } } },
      "put": { "summary": "Update a file by ID", "tags": ["File"], "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } } ], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FileInput" } } } }, "responses": { "200": { "description": "File updated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/File" } } } }, "400": { "description": "Invalid input" }, "404": { "description": "File not found" } } },
      "delete": { "summary": "Delete a file by ID", "tags": ["File"], "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } } ], "responses": { "204": { "description": "File deleted" }, "404": { "description": "File not found" } } }
    },
...
    "/api/legal_acts": {
      "get": { "summary": "Get all legal acts", "tags": ["LegalAct"], "responses": { "200": { "description": "List of legal acts", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/LegalAct" } } } } } } },
      "post": { "summary": "Create a new legal act", "tags": ["LegalAct"], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LegalActInput" } } } }, "responses": { "201": { "description": "Legal act created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LegalAct" } } } }, "400": { "description": "Invalid input" } } }
    },
    "/api/legal_acts/{id}": {
      "get": { "summary": "Get a legal act by ID", "tags": ["LegalAct"], "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } } ], "responses": { "200": { "description": "Legal act found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LegalAct" } } } }, "404": { "description": "Legal act not found" } } },
      "put": { "summary": "Update a legal act by ID", "tags": ["LegalAct"], "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } } ], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LegalActInput" } } } }, "responses": { "200": { "description": "Legal act updated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LegalAct" } } } }, "400": { "description": "Invalid input" }, "404": { "description": "Legal act not found" } } },
      "delete": { "summary": "Delete a legal act by ID", "tags": ["LegalAct"], "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } } ], "responses": { "204": { "description": "Legal act deleted" }, "404": { "description": "Legal act not found" } } }
    },
...
    "/api/emails": {
      "get": { "summary": "Get all emails", "tags": ["Email"], "responses": { "200": { "description": "List of emails", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Email" } } } } } } },
      "post": { "summary": "Create a new email", "tags": ["Email"], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EmailInput" } } } }, "responses": { "201": { "description": "Email created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Email" } } } }, "400": { "description": "Invalid input" } } }
    },
    "/api/emails/{id}": {
      "get": { "summary": "Get an email by ID", "tags": ["Email"], "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } } ], "responses": { "200": { "description": "Email found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Email" } } } }, "404": { "description": "Email not found" } } },
      "put": { "summary": "Update an email by ID", "tags": ["Email"], "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } } ], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EmailInput" } } } }, "responses": { "200": { "description": "Email updated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Email" } } } }, "400": { "description": "Invalid input" }, "404": { "description": "Email not found" } } },
      "delete": { "summary": "Delete an email by ID", "tags": ["Email"], "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } } ], "responses": { "204": { "description": "Email deleted" }, "404": { "description": "Email not found" } } }
    },
...
    "/api/file/export": {
      "post": {
        "summary": "Ajouter un export de fichier (nom et chemins)",
        "description": "Ajoute un enregistrement dans la table 'export' avec le nom du fichier (sans extension) et un tableau de chemins (CSV, XLSX, etc).",
        "tags": ["Export"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": { "type": "string", "description": "Nom du fichier exporté (sans extension)" },
                  "path": { "type": "array", "items": { "type": "string" }, "description": "Tableau des chemins des fichiers exportés (CSV, XLSX, etc)" }
                },
                "required": ["file", "path"],
                "example": { "file": "export", "path": ["/public/file/export.csv", "/public/file/export.xlsx"] }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Export enregistré avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "integer" },
                    "file": { "type": "string" },
                    "path": { "type": "array", "items": { "type": "string" } },
                    "created_at": { "type": "string", "format": "date-time" }
                  }
                }
              }
            }
          },
          "400": { "description": "Requête invalide (file ou path manquant)" },
          "500": { "description": "Erreur serveur" }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Establishment": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid", "description": "Unique identifier (UUID)" },
          "siret": { "type": "string", "description": "SIRET number (max 20 characters)" },
          "is_head_office": { "type": "boolean", "description": "Indicates if this is the head office" },
          "creation_date": { "type": "string", "format": "date", "nullable": true, "description": "Date of establishment creation" },
          "status": { "type": "string", "nullable": true, "description": "Establishment status" },
          "address_id": { "type": "string", "format": "uuid", "nullable": true, "description": "Reference to address" },
          "company_id": { "type": "string", "format": "uuid", "description": "Reference to company" },
          "address": { "$ref": "#/components/schemas/Address", "nullable": true, "description": "Associated address" },
          "company": { "$ref": "#/components/schemas/Company", "description": "Associated company" }
        },
        "required": ["id", "siret", "is_head_office", "company_id"]
      },
      "EstablishmentInput": {
        "type": "object",
        "properties": {
          "siret": { "type": "string", "description": "SIRET number (max 20 characters)" },
          "is_head_office": { "type": "boolean", "description": "Indicates if this is the head office" },
          "creation_date": { "type": "string", "format": "date", "nullable": true, "description": "Date of establishment creation" },
          "status": { "type": "string", "nullable": true, "description": "Establishment status" },
          "address_id": { "type": "string", "format": "uuid", "nullable": true, "description": "Reference to address" },
          "company_id": { "type": "string", "format": "uuid", "description": "Reference to company" }
        },
        "required": ["siret", "is_head_office", "company_id"]
      },
      "Officer": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid", "description": "Unique identifier (UUID)" },
          "first_name": { "type": "string", "nullable": true, "description": "Officer's first name" },
          "last_name": { "type": "string", "nullable": true, "description": "Officer's last name" },
          "role": { "type": "string", "nullable": true, "description": "Officer's role (e.g., CEO, CFO)" },
          "start_date": { "type": "string", "format": "date", "nullable": true, "description": "Start date of role" },
          "end_date": { "type": "string", "format": "date", "nullable": true, "description": "End date of role" },
          "registration_number": { "type": "string", "nullable": true, "description": "Registration number if officer is a company (max 20 characters)" },
          "is_company": { "type": "boolean", "nullable": true, "description": "Indicates if the officer is a company" },
          "company_id": { "type": "string", "format": "uuid", "description": "Reference to company" },
          "company": { "$ref": "#/components/schemas/Company", "description": "Associated company" }
        },
        "required": ["id", "company_id"]
      },
      "OfficerInput": {
        "type": "object",
        "properties": {
          "first_name": { "type": "string", "nullable": true, "description": "Officer's first name" },
          "last_name": { "type": "string", "nullable": true, "description": "Officer's last name" },
          "role": { "type": "string", "nullable": true, "description": "Officer's role" },
          "start_date": { "type": "string", "format": "date", "nullable": true, "description": "Start date of role" },
          "end_date": { "type": "string", "format": "date", "nullable": true, "description": "End date of role" },
          "registration_number": { "type": "string", "nullable": true, "description": "Registration number if officer is a company" },
          "is_company": { "type": "boolean", "nullable": true, "description": "Indicates if the officer is a company" },
          "company_id": { "type": "string", "format": "uuid", "description": "Reference to company" }
        },
        "required": ["company_id"]
      },
...
      "LegalAct": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid", "description": "Unique identifier (UUID)" },
          "date": { "type": "string", "format": "date", "nullable": true, "description": "Date of legal act" },
          "type": { "type": "string", "nullable": true, "description": "Type of legal act" },
          "description": { "type": "string", "nullable": true, "description": "Legal act description" },
          "company_id": { "type": "string", "format": "uuid", "description": "Reference to company" },
          "company": { "$ref": "#/components/schemas/Company", "description": "Associated company" }
        },
        "required": ["id", "company_id"]
      },
      "LegalActInput": {
        "type": "object",
        "properties": {
          "date": { "type": "string", "format": "date", "nullable": true, "description": "Date of legal act" },
          "type": { "type": "string", "nullable": true, "description": "Type of legal act" },
          "description": { "type": "string", "nullable": true, "description": "Legal act description" },
          "company_id": { "type": "string", "format": "uuid", "description": "Reference to company" }
        },
        "required": ["company_id"]
      },
...
      "Email": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid", "description": "Unique identifier (UUID)" },
          "value": { "type": "string", "description": "Email address" },
          "type": { "type": "string", "enum": ["generic", "personal"], "description": "Email type" },
          "test_result": { "type": "string", "nullable": true, "description": "Result of email validation test" },
          "test_date": { "type": "string", "format": "date", "nullable": true, "description": "Date of email validation test" },
          "company_id": { "type": "string", "format": "uuid", "description": "Reference to company" },
          "company": { "$ref": "#/components/schemas/Company", "description": "Associated company" }
        },
        "required": ["id", "value", "type", "company_id"]
      },
      "EmailInput": {
        "type": "object",
        "properties": {
          "value": { "type": "string", "description": "Email address" },
          "type": { "type": "string", "enum": ["generic", "personal"], "description": "Email type" },
          "test_result": { "type": "string", "nullable": true, "description": "Result of email validation test" },
          "test_date": { "type": "string", "format": "date", "nullable": true, "description": "Date of email validation test" },
          "company_id": { "type": "string", "format": "uuid", "description": "Reference to company" }
        },
        "required": ["value", "type", "company_id"]
      },
      "Address": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid", "description": "Unique identifier (UUID)" },
          "street": { "type": "string", "nullable": true, "description": "Street address" },
          "postal_code": { "type": "string", "nullable": true, "description": "Postal code (max 10 characters)" },
          "city": { "type": "string", "nullable": true, "description": "City" },
          "country_code": { "type": "string", "nullable": true, "description": "Country code (max 5 characters)" },
          "lat": { "type": "number", "nullable": true, "description": "Latitude" },
          "lng": { "type": "number", "nullable": true, "description": "Longitude" }
        },
        "required": ["id"]
      },
...
      "File": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid", "description": "Unique identifier (UUID)" },
          "file_name": { "type": "string", "nullable": true, "description": "Name of the file" },
          "upload_date": { "type": "string", "format": "date-time", "nullable": true, "description": "File upload date" },
          "row_count": { "type": "integer", "nullable": true, "description": "Number of rows in the file" },
          "enriched_row_count": { "type": "integer", "nullable": true, "description": "Number of enriched rows" },
          "status": {
            "type": "string",
            "enum": ["uploaded", "processing", "completed", "error"],
            "nullable": true,
            "description": "File processing status"
          },
          "first_name": { "type": "string", "nullable": true, "description": "First name from file data" },
          "last_name": { "type": "string", "nullable": true, "description": "Last name from file data" },
          "gender": { "type": "string", "nullable": true, "description": "Gender from file data" },
          "title": { "type": "string", "nullable": true, "description": "Job title from file data" },
          "email": { "type": "string", "nullable": true, "description": "Email from file data" },
          "phone": { "type": "string", "nullable": true, "description": "Phone number from file data" },
          "all_phones": { "type": "string", "nullable": true, "description": "All phone numbers as text" },
          "company_name": { "type": "string", "nullable": true, "description": "Company name from file data" },
          "company_cleaned_name": { "type": "string", "nullable": true, "description": "Cleaned company name" },
          "company_website": { "type": "string", "nullable": true, "description": "Company website URL" },
          "company_domain": { "type": "string", "nullable": true, "description": "Company domain" },
          "linkedin_profile_url": { "type": "string", "nullable": true, "description": "LinkedIn profile URL" },
          "company_linkedin_id_url": { "type": "string", "nullable": true, "description": "Company LinkedIn ID URL" },
          "location": { "type": "string", "nullable": true, "description": "Location from file data" },
          "company_location": { "type": "string", "nullable": true, "description": "Company location" },
          "company_industry": { "type": "string", "nullable": true, "description": "Company industry" },
          "employee_count": { "type": "string", "nullable": true, "description": "Employee count as text" },
          "employee_range": { "type": "string", "nullable": true, "description": "Employee count range" },
          "revenue": { "type": "string", "nullable": true, "description": "Revenue as text" },
          "headquarters": { "type": "string", "nullable": true, "description": "Company headquarters location" },
          "year_founded": { "type": "string", "nullable": true, "description": "Year company was founded" },


          "linkedin_headline": { "type": "string", "nullable": true, "description": "LinkedIn headline" },
          "linkedin_connections_count": { "type": "integer", "nullable": true, "description": "Number of LinkedIn connections" },
          "start_date_job": { "type": "string", "nullable": true, "description": "Job start date" },
          "start_date_company": { "type": "string", "nullable": true, "description": "Company start date" },
          "years_in_position": { "type": "integer", "nullable": true, "description": "Years in current position" },
          "months_in_position": { "type": "integer", "nullable": true, "description": "Months in current position" },
          "years_in_company": { "type": "integer", "nullable": true, "description": "Years in company" },
          "months_in_company": { "type": "integer", "nullable": true, "description": "Months in company" },
          "current_positions_count": { "type": "integer", "nullable": true, "description": "Number of current positions" },
          "title_description": { "type": "string", "nullable": true, "description": "Job title description" },
          "summary": { "type": "string", "nullable": true, "description": "Profile summary" },
          "full_name": { "type": "string", "nullable": true, "description": "Full name from file data" },
          "sales_navigator_profile_url": { "type": "string", "nullable": true, "description": "Sales Navigator profile URL" },
          "linkedin_id_url": { "type": "string", "nullable": true, "description": "LinkedIn ID URL" },
          "profile_image_url": { "type": "string", "nullable": true, "description": "Profile image URL" },
          "company_profile_picture": { "type": "string", "nullable": true, "description": "Company profile picture URL" },
          "company_description": { "type": "string", "nullable": true, "description": "Company description" },
          "company_linkedin": { "type": "string", "nullable": true, "description": "Company LinkedIn URL" },
          "connection_degree": { "type": "string", "nullable": true, "description": "LinkedIn connection degree" },
          "is_premium_linkedin": { "type": "boolean", "nullable": true, "description": "Indicates if LinkedIn profile is premium" },
          "is_open_profile_linkedin": { "type": "boolean", "nullable": true, "description": "Indicates if LinkedIn profile is open" },
          "is_open_to_work_linkedin": { "type": "boolean", "nullable": true, "description": "Indicates if open to work on LinkedIn" },
          "email_alt": { "type": "string", "nullable": true, "description": "Alternative email" },
          "phone_alt": { "type": "string", "nullable": true, "description": "Alternative phone number" },
          "all_phones_alt": { "type": "string", "nullable": true, "description": "All alternative phone numbers as text" },
          "company_id": { "type": "string", "format": "uuid", "nullable": true, "description": "Reference to company" },
          "company": { "$ref": "#/components/schemas/Company", "nullable": true, "description": "Associated company" }
        },
        "required": ["id"]
      },
...
    }
  },
...
}